chore: add eslint react-refresh plugin and separate useToast hook
- Add eslint-plugin-react-refresh to dependencies
- Configure react-refresh plugin in eslint config
- Move useToast hook to a separate file for better fast refresh support
