+++ eslint.config.mjs
@@ -1,5 +1,6 @@
 // @ts-check
 import reactHooks from 'eslint-plugin-react-hooks';
+import reactRefresh from 'eslint-plugin-react-refresh';
 import eslint from '@eslint/js';
 import jest from 'eslint-plugin-jest';
 import eslintPluginUnicorn from 'eslint-plugin-unicorn';
@@ -40,6 +41,7 @@ export default [
       'react-hooks': reactHooks,
       'unused-imports': unusedImports,
       unicorn: eslintPluginUnicorn,
+      'react-refresh': reactRefresh,
     },
     rules: {
       '@typescript-eslint/no-unsafe-function-type': 0,
@@ -70,6 +72,7 @@ export default [
       'unicorn/no-negated-condition': 'error',
       'unicorn/prefer-number-properties': 'error',
       'unused-imports/no-unused-imports': 'error',
+      'react-refresh/only-export-components': 'warn',
     },
   },
   {
