+++ package.json
@@ -120,6 +120,7 @@
     "eslint": "^9.12.0",
     "eslint-plugin-jest": "^28.8.3",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "^0.4.13",
     "eslint-plugin-unicorn": "^56.0.0",
     "eslint-plugin-unused-imports": "^4.1.4",
     "jest": "^29.7.0",
